+++ packages/web3-eth-contract/src/contract.ts
@@ -36,6 +36,8 @@ import {
 	ContractEvent,
 	ContractEvents,
 	ContractMethod,
+	ContractMethodInputParameters,
+	ContractMethodOutputParameters,
 	encodeEventSignature,
 	encodeFunctionSignature,
 	FilterAbis,
@@ -102,7 +104,6 @@ import {
 	isWeb3ContractContext,
 } from './utils';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
 type ContractBoundMethod<
 	Abi extends AbiFunctionFragment,
 	Method extends ContractMethod<Abi> = ContractMethod<Abi>,
@@ -111,6 +112,26 @@ type ContractBoundMethod<
 ) => Method['Abi']['stateMutability'] extends 'payable' | 'pure'
 	? PayableMethodObject<Method['Inputs'], Method['Outputs']>
 	: NonPayableMethodObject<Method['Inputs'], Method['Outputs']>;
+
+export type ContractOverloadedMethodInputs<AbiArr extends ReadonlyArray<unknown>> = NonNullable<
+	AbiArr extends readonly []
+		? undefined
+		: AbiArr extends readonly [infer A, ...infer R]
+		? A extends AbiFunctionFragment
+			? ContractMethodInputParameters<A['inputs']> | ContractOverloadedMethodInputs<R>
+			: undefined
+		: undefined
+>;
+
+export type ContractOverloadedMethodOutputs<AbiArr extends ReadonlyArray<unknown>> = NonNullable<
+	AbiArr extends readonly []
+		? undefined
+		: AbiArr extends readonly [infer A, ...infer R]
+		? A extends AbiFunctionFragment
+			? ContractMethodOutputParameters<A['outputs']> | ContractOverloadedMethodOutputs<R>
+			: undefined
+		: undefined
+>;
 
 // To avoid circular dependency between types and encoding, declared these types here.
 export type ContractMethodsInterface<Abi extends ContractAbi> = {
@@ -912,25 +933,20 @@ export class Contract<Abi extends ContractAbi>
 					abi.constant;
 
 				abi.payable = abi.stateMutability === 'payable' ?? abi.payable;
-
-				const contractMethod = this._createContractMethod(abi, errorsAbi);
-
 				this._overloadedMethodAbis.set(abi.name, [
 					...(this._overloadedMethodAbis.get(abi.name) ?? []),
 					abi,
 				]);
 
-				if (methodName in this._functions) {
-					this._functions[methodName] = {
-						signature: methodSignature,
-						method: contractMethod,
-					};
-				} else {
-					this._functions[methodName] = {
-						signature: methodSignature,
-						method: contractMethod,
-					};
-				}
+				const contractMethod = this._createContractMethod(
+					this._overloadedMethodAbis.get(abi.name) ?? [],
+					errorsAbi,
+				);
+
+				this._functions[methodName] = {
+					signature: methodSignature,
+					method: contractMethod,
+				};
 
 				// We don't know a particular type of the Abi method so can't type check
 				this._methods[abi.name as keyof ContractMethodsInterface<Abi>] = this._functions[
@@ -979,10 +995,11 @@ export class Contract<Abi extends ContractAbi>
 			);
 		}
 	}
-	private _createContractMethod<T extends AbiFunctionFragment, E extends AbiErrorFragment>(
-		abi: T,
+	private _createContractMethod<T extends AbiFunctionFragment[], E extends AbiErrorFragment>(
+		abiArr: T,
 		errorsAbis: E[],
-	): ContractBoundMethod<T> {
+	): ContractBoundMethod<T[0]> {
+		const abi = abiArr[abiArr.length - 1];
 		return (...params: unknown[]) => {
 			let abiParams!: Array<unknown>;
 			const abis = this._overloadedMethodAbis.get(abi.name) ?? [];
@@ -1036,8 +1053,8 @@ export class Contract<Abi extends ContractAbi>
 						}),
 					encodeABI: () => encodeMethodABI(methodAbi, abiParams),
 				} as unknown as PayableMethodObject<
-					ContractMethod<T>['Inputs'],
-					ContractMethod<T>['Outputs']
+					ContractOverloadedMethodInputs<T>,
+					ContractOverloadedMethodOutputs<T>
 				>;
 			}
 			return {
@@ -1058,8 +1075,8 @@ export class Contract<Abi extends ContractAbi>
 					}),
 				encodeABI: () => encodeMethodABI(methodAbi, abiParams),
 			} as unknown as NonPayableMethodObject<
-				ContractMethod<T>['Inputs'],
-				ContractMethod<T>['Outputs']
+				ContractOverloadedMethodInputs<T>,
+				ContractOverloadedMethodOutputs<T>
 			>;
 		};
 	}